chore(deps): upgrade to astro 6 + @astrojs/{mdx,react} 5
Closes the second moderate CVE (GHSA-j687-52p2-xcff: XSS in define:vars).
Site doesn't use define:vars so the CVE was never exploitable here, but
bumping is good hygiene now that React landed and the redesign work
benefits from current-major Astro.

Versions: astro 5.18.1 → 6.2.1, @astrojs/mdx 4.3.13 → 5.0.4,
@astrojs/react 4.4.2 → 5.0.4. @astrojs/sitemap 3.7.2 stays (loose-coupled,
no peer-dep on astro).

Audit: 2 moderate → 0 vulnerabilities. ✅

Breaking-change migration:

- Move src/content/config.ts → src/content.config.ts (v6 file location).
- Replace legacy `type: 'content'` collection with `loader: glob({ pattern,
  base })` from `astro/loaders`. v6 removed legacy backwards-compat for
  the original Content Collections API.
- Replace `entry.render()` with `render(entry)` from `astro:content` in
  src/pages/vision/index.astro. The legacy method was removed.
- Update tests/content-collections.spec.ts to assert the new paths and APIs.

Verification:
- npm run build: green, 16 pages, 1.85s.
- npm run postbuild: 0 errors (CSP + SRI hashes regenerated for new chunks).
- node --test tests/tweaks-state.test.mts: 12/12 passing.
- node --test tests/build-config.test.mjs: 4/4 passing.
- npx playwright test tests/styling-a11y.spec.ts: 2/2 passing.
- npm audit: 0 vulnerabilities.

Items checked but NOT used by this site (so no migration needed):
- ViewTransitions, Astro.glob(), ASSETS_PREFIX, ClientRouter, getStaticPaths,
  CommonJS config files, all adapter APIs (output: "static").

Deferred (deprecated only, not removed):
- `import { z } from 'astro:content'` works in 6.x with a deprecation;
  v7 will require `from 'astro/zod'`. Single-line edit when needed.
