+++ package.json
@@ -62,10 +62,10 @@
     "prepare": "husky"
   },
   "dependencies": {
-    "@astrojs/mdx": "4.3.13",
-    "@astrojs/react": "4.4.2",
+    "@astrojs/mdx": "5.0.4",
+    "@astrojs/react": "5.0.4",
     "@astrojs/sitemap": "3.7.2",
-    "astro": "5.18.1",
+    "astro": "6.2.1",
     "react": "19.2.5",
     "react-dom": "19.2.5",
     "sharp": "0.34.5"
