--- src/content/config.ts
@@ -1,18 +1,0 @@
-import { defineCollection, z } from 'astro:content';
-
-const pagesCollection = defineCollection({
-  type: 'content',
-  schema: z.object({
-    title: z.string(),
-    description: z.string(),
-    hero: z.object({
-      title: z.string(),
-      subtitle: z.string(),
-      missionText: z.string(),
-    }).optional(),
-  }),
-});
-
-export const collections = {
-  pages: pagesCollection,
-};
