+++ src/pages/vision/index.astro
@@ -1,5 +1,5 @@
 ---
-import { getEntry } from 'astro:content';
+import { getEntry, render } from 'astro:content';
 import Base from '../../layouts/BaseLayout.astro';
 import Header from '../../components/Header.astro';
 import Footer from '../../components/Footer.astro';
@@ -10,7 +10,7 @@ if (!entry) {
   throw new Error('Vision page content not found');
 }
 
-const { Content } = await entry.render();
+const { Content } = await render(entry);
 const { title, description } = entry.data;
 ---
